pyproject.toml
[project]
name = "asciifer"
version = "0.1.0"
description = ""
authors = [
    {name = "Refffy",email = "gvertyes124@gmail.com"}
]
readme = "README.md"
requires-python = ">=3.11"
dependencies = [
    "pillow (>=11.3.0,<12.0.0)",
    "numpy (>=2.3.4,<3.0.0)"
]

[tool.poetry]
packages = [{include = "asciifer", from = "src"}]


[build-system]
requires = ["poetry-core>=2.0.0,<3.0.0"]
build-backend = "poetry.core.masonry.api"
